[SPOT-256][FEAT] 테마 별 조회 위한 Request DTO 재정의
