+++ src/main/java/com/example/spot/web/dto/search/SearchRequestStudyDTO.java
@@ -1,0 +1,25 @@
+package com.example.spot.web.dto.search;
+
+import com.example.spot.domain.enums.Gender;
+import com.example.spot.domain.enums.ThemeType;
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.AssertTrue;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@Setter
+@SuperBuilder
+@NoArgsConstructor
+public class SearchRequestStudyDTO extends BaseSearchRequestStudyDTO {
+
+}
+
